# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Oil \u0026 Gas Service Booking API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
...
        "/booking-services": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking-services"
                ],
                "summary": "Получить все услуги в бронированиях",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/oil-gas-service-booking_internal_models.BookingService"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking-services"
                ],
                "summary": "Создать услугу в бронировании",
                "parameters": [
                    {
                        "description": "Данные услуги в бронировании",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http-server_handlers.BookingServiceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/oil-gas-service-booking_internal_models.BookingService"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/booking-services/{id}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking-services"
                ],
                "summary": "Получить услугу в бронировании по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "BookingService ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/oil-gas-service-booking_internal_models.BookingService"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking-services"
                ],
                "summary": "Обновить услугу в бронировании",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "BookingService ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/oil-gas-service-booking_internal_models.BookingService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/oil-gas-service-booking_internal_models.BookingService"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "tags": [
                    "booking-services"
                ],
                "summary": "Удалить услугу из бронирования",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "BookingService ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/bookings/{booking_id}/services": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking-services"
                ],
                "summary": "Получить услуги по ID бронирования",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Booking ID",
                        "name": "booking_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/oil-gas-service-booking_internal_models.BookingService"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_http-server_handlers.BookingServiceRequest": {
            "type": "object",
            "properties": {
                "booking_id": {
                    "type": "integer"
                },
                "company_service_id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                }
            }
        },
...
        "oil-gas-service-booking_internal_models.Booking": {
            "type": "object",
            "properties": {
                "bookingID": {
                    "type": "integer"
                },
                "bookingServices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oil-gas-service-booking_internal_models.BookingService"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/oil-gas-service-booking_internal_models.User"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "oil-gas-service-booking_internal_models.BookingService": {
            "type": "object",
            "properties": {
                "booking": {
                    "$ref": "#/definitions/oil-gas-service-booking_internal_models.Booking"
                },
                "bookingID": {
                    "type": "integer"
                },
                "bookingServiceID": {
                    "type": "integer"
                },
                "companyService": {
                    "$ref": "#/definitions/oil-gas-service-booking_internal_models.CompanyService"
                },
                "companyServiceID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "oil-gas-service-booking_internal_models.Company": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "companyID": {
                    "type": "integer"
                },
                "companyServices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oil-gas-service-booking_internal_models.CompanyService"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/oil-gas-service-booking_internal_models.User"
                },
                "userID": {
                    "description": "владелец",
                    "type": "integer"
                }
            }
        },
        "oil-gas-service-booking_internal_models.CompanyService": {
            "type": "object",
            "properties": {
                "bookingServices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oil-gas-service-booking_internal_models.BookingService"
                    }
                },
                "company": {
                    "$ref": "#/definitions/oil-gas-service-booking_internal_models.Company"
                },
                "companyID": {
                    "type": "integer"
                },
                "companyServiceID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "price": {
                    "description": "цена конкретной услуги у компании",
                    "type": "number"
                },
                "service": {
                    "$ref": "#/definitions/oil-gas-service-booking_internal_models.Service"
                },
                "serviceID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "oil-gas-service-booking_internal_models.Service": {
            "type": "object",
            "properties": {
                "companyServices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oil-gas-service-booking_internal_models.CompanyService"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "serviceID": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "oil-gas-service-booking_internal_models.User": {
            "type": "object",
            "properties": {
                "bookings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oil-gas-service-booking_internal_models.Booking"
                    }
                },
                "companies": {
                    "description": "Добавляем обратную связь",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/oil-gas-service-booking_internal_models.Company"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}